Split image composer persistence hook
